* feat/factory_fw/upload_warn_cloud:
  feat: [factory_fw] end task from cloud, ack back ended state
  cherry-pic: fix: [factory_fw] initialize the app in CORE1 and allocate interrupt resources wisely
  fix: [factory_fw] increase nvs partition size to 128k; WIP upload warn to cloud

+++ examples/factory_firmware/main/app/app_mqtt_client.h
@@ -2,6 +2,8 @@
 
 #pragma once
 
+#include <stdint.h>
+
 #include "esp_err.h"
 #include "cJSON.h"
 
@@ -13,9 +15,9 @@ esp_err_t app_mqtt_client_init(void);
 
 esp_err_t app_mqtt_client_report_tasklist_ack(char *request_id, cJSON *task_settings_node);
 
-esp_err_t app_mqtt_client_report_tasklist_status(int tasklist_id, int tasklist_status_num);
+esp_err_t app_mqtt_client_report_tasklist_status(intmax_t tasklist_id, int tasklist_status_num);
 
-esp_err_t app_mqtt_client_report_warn_event(int tasklist_id, char *tasklist_name, int warn_type);
+esp_err_t app_mqtt_client_report_warn_event(intmax_t tasklist_id, char *tasklist_name, int warn_type);
 
 
 #ifdef __cplusplus
